Move cast_vote_and_cascade_follow
